refactor: move blueprint hooks from ChatOptions to dedicated module and update UI
